feat: ArticleのuseContainerをやめて全てindex.tsxに処理をまとめる
- useContainer.tsの処理をindex.tsxに統合
- GraphQLクエリとロジックを一箇所に集約
- 関連するモックファイルとテストファイルのリネーム
- 不要なファイルの削除
