+++ mobile/public/mockServiceWorker.js
@@ -1,4 +1,4 @@
-/* eslint-disable */
+ 
 /* tslint:disable */
 
 /**
