+++ mobile/src/features/Articles/index.test.tsx
@@ -1,8 +1,8 @@
 import { composeStories } from "@storybook/react";
 import { server } from "@/libs/test/server";
 import { render, screen } from "@/libs/test/testing-library";
+import { handlers } from "./index.mocks";
 import * as stories from "./index.stories";
-import { handlers } from "./useContainer.mocks";
 
 const { Primary } = composeStories(stories);
